- [AI] Adjusted obstacle avoidance to have box extend and use agent radius - [Graphics] Added Drawing Screen Triangle support - Added new project Flocking to test flocking behavior
git-tfs-id: [https://dev.azure.com/timjustin/]$/Angazi;C367

+++ VGP337/NEAT/GameState.cpp
@@ -113,7 +113,7 @@ namespace
 				{
 					if (birds[activeBirdCount].fitness > bestFitness)
 						bestFitness = birds[activeBirdCount].fitness;
-					g.fitness = birds[activeBirdCount++].fitness;
+					g.fitness = static_cast<size_t>(birds[activeBirdCount++].fitness);
 				}
 
 			population->NewGeneration();
